% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxSE.R
\name{approxSE}
\alias{approxSE}
\title{Approximate standard errors using the empirical information matrix}
\usage{
approxSE(theta, l, t, z, m)
}
\description{
Approximate standard errors using the empirical information matrix
}
\keyword{internal}

