% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSE.R
\name{bootSE}
\alias{bootSE}
\title{Standard errors via bootstrap for an \code{mjoint} object}
\usage{
bootSE(object, nboot = 100, ci = 0.95, use.mle = TRUE, verbose = FALSE,
  control = list(), progress = TRUE, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{mjoint} for a joint model
of time-to-event and multivariate longitudinal data.}

\item{nboot}{the number of bootstrap samples. Default is \code{nboot=100}.}

\item{ci}{the confidence interval to be estimated using the
percentile-method. Default is \code{ci=0.95} for a 95\% confidence
interval.}

\item{use.mle}{logical: should the algorithm use the maximiser from the
converged model in \code{object} as initial values for coefficients in each
bootstrap iteration. Default is \code{use.mle=TRUE}.}

\item{verbose}{logical: if \code{TRUE}, the parameter estimates and other
convergence statistics are value are printed at each iteration of the MCEM
algorithm. Default is \code{FALSE}.}

\item{control}{a list of control values with components: \describe{

  \item{\code{nMC}}{integer: the initial number of Monte Carlo samples to be
  used for integration in the early phase of the MCEM. Default is
  \code{nMC=100}.}

  \item{\code{nMCscale}}{integer: the scale factor for the increase in Monte
  Carlo size when Monte Carlo has not reduced from the previous iteration.
  Default is \code{nMCscale=3}.}

  \item{\code{nMCmax}}{integer: the maximum number of Monte Carlo samples
  that the algorithm is allowed to reach. Default is \code{nMCmax=20000}.}

  \item{\code{earlyPhase}}{integer: the number of iterations for early phase
  of the optimization algorithm. It is computationally inefficient to use a
  large number of Monte Carlo samples early on until one is approximately
  near the maximum likelihood estimate. Default is \code{earlyPhase=30}.}

  \item{\code{mcmaxIter}}{integer: the maximum number of MCEM algorithm
  iterations allowed. Default is \code{mcmaxIter=200}.}

  \item{\code{convCrit}}{character string: the convergence criterion to be
  used. See \strong{Details}.}

  \item{\code{approxInfo}}{logical: should the information matrix in the
  Newton-Raphson step for the \eqn{\gamma} parameters be calculated using an
  approximation based on the empirical information matrix? If it is used,
  then the step-length is adjusted by a nominal scaling parameter of 0.5 in
  order to increase the probability that the \emph{Q}-function in the MCEM
  algorithm is increased at each iteration (McLachlan and Krishnan, 2008).
  Default is \code{FALSE}, which corresponds to an exact calculation.}

  \item{\code{tol0}}{numeric: tolerance value for convergence in the
  parameters; see \strong{Details}. Default is \code{5e-03}.}

  \item{\code{tol1}}{numeric: tolerance value for convergence in the
  parameters; see \strong{Details}. Default is \code{1e-03}.}

  \item{\code{tol2}}{numeric: tolerance value for convergence in the
  parameters; see \strong{Details}. Default is \code{5e-03}.}

  \item{\code{rav}}{numeric: threshold when using \code{convCrit='sas'} that
  applies absolute change (when <\code{rav}) or relative change (when
  \eqn{\geq}\code{rav}) criterion; see \strong{Details}. Default is
  \code{0.1}, which is an order of magnitude higher than the SAS
  implementation.}

  }}

\item{progress}{logical: should a progress bar be shown on the console to
indicate the percentage of bootstrap iterations completed? Default is
\code{progress=TRUE}.}

\item{...}{options passed to the \code{control} argument.}
}
\value{
An object of class \code{bootSE}.
}
\description{
This function takes a model fit from an \code{mjoint} object and calculates
standard errors and confidence intervals for the main longitudinal and
survival coefficient parameters, including the latent association parameters,
using bootstrapping (Efron and Tibshirani, 2000).
}
\details{
Standard errors and confidence intervals are obtained by repeated
  fitting of the requisite joint model to bootstrap samples of the original
  longitudinal and time-to-event data. Note that bootstrap is done by
  sampling subjects, not individual records.
}
\note{
This function is computationally intensive. A dynamic progress bar is
  displayed showing the percentage of bootstrap models fitted.

  Due to random sampling, an \code{mjoint} model fitted to some bootstrap
  samples may not converge within the specified control parameter settings.
  The \code{bootSE} code discards any models that failed to converge when
  calculating the standard errors and confidence intervals. If a large
  proportion of models have failed to converge, it is likely that it will
  need to be refitted with changes to the \code{control} arguments.
}
\examples{
\dontrun{
# Fit a joint model with bivariate longitudinal outcomes

data(heart.valve)
hvd <- heart.valve[!is.na(heart.valve$log.grad) & !is.na(heart.valve$log.lvmi), ]

fit2 <- mjoint(
    formLongFixed = list("grad" = log.grad ~ time + sex + hs,
                         "lvmi" = log.lvmi ~ time + sex),
    formLongRandom = list("grad" = ~ 1 | num,
                          "lvmi" = ~ time | num),
    formSurv = Surv(fuyrs, status) ~ age,
    data = list(hvd, hvd),
    inits = list("gamma" = c(0.11, 1.51, 0.80)),
    timeVar = "time",
    verbose = TRUE)

fit2.boot <- bootSE(fit1, 50, use.mle = TRUE, control = list(
    earlyPhase = 25, convCrit = "either",
    tol0 = 6e-03, tol2 = 6e-03, mcmaxIter = 60))
}
}
\author{
Graeme L. Hickey (\email{graeme.hickey@liverpool.ac.uk})
}
\references{
Efron B, Tibshirani R. \emph{An Introduction to the Bootstrap.} 2000; Boca
Raton, FL: Chapman & Hall/CRC.
}
\seealso{
\code{\link{mjoint}} for approximate standard errors.
}
\keyword{methods}
\keyword{multivariate}
\keyword{survival}

