% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convMonitor.R
\name{convMonitor}
\alias{convMonitor}
\title{Evaluate the convergence of the MCEM algorithm}
\usage{
convMonitor(theta, theta.new, log.lik, log.lik.new, con, verbose)
}
\description{
Evaluate the convergence of the MCEM algorithm
}
\keyword{internal}

