% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_nearPD.R
\name{fast_nearPD}
\alias{fast_nearPD}
\title{Fast calculation of a near postive-definite matrix}
\usage{
fast_nearPD(M, eig.tol = 1e-06, conv.tol = 1e-07, posd.tol = 1e-08,
  maxits = 100)
}
\description{
Code is from JM package (v. 1.4-5) by Dimitris Rizopoulos
Copied here due to function being non-exported by JM package, and also because
the Matrix package version is an order of magnitude slower.
This code is originally based on function nearcor() submitted to R-help by
Jens Oehlschlagel on 2007-07-13, and function posdefify() from package 'sfsmisc'
}
\keyword{internal}

