% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gammaUpdate}
\alias{gammaUpdate}
\title{Newton-Raphson updates of gamma (E-step and M-step) using an exact observed
information calculation}
\usage{
gammaUpdate(b_, z_, w_, pb_, haz, v_, h_, K, q, nev)
}
\description{
Newton-Raphson updates of gamma (E-step and M-step) using an exact observed
information calculation
}
\keyword{internal}

