% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gammaUpdate_approx}
\alias{gammaUpdate_approx}
\title{Newton-Raphson updates of gamma (E-step and M-step) using an empirical
infomration matrix approximation}
\usage{
gammaUpdate_approx(b_, z_, w_, pb_, haz, v_, h_, K, q, nev)
}
\description{
Newton-Raphson updates of gamma (E-step and M-step) using an empirical
infomration matrix approximation
}
\keyword{internal}

