% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{AIC.mjoint}
\alias{AIC.mjoint}
\alias{BIC.mjoint}
\title{Extract information criteria from an \code{mjoint} object}
\usage{
\method{AIC}{mjoint}(object, ..., k = 2)

\method{BIC}{mjoint}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mjoint}.}

\item{...}{Not used.}

\item{k}{The penalty applied to the AIC. Defaults to 2.}
}
\value{
Returns a single number with the AIC or the BIC value for the \code{object} passed to the function.
}
\description{
Extract the Akaike Information Criterion (AIC) and the Bayesian Information Criterion (BIC) from an \code{mjoint} object.
}
\note{
The penalty value for the BIC is the natural logarithm of the number of individuals.
}
\examples{
\dontrun{
# Fit a joint model with bivariate longitudinal outcomes

data(heart.valve)
hvd <- heart.valve[!is.na(heart.valve$log.grad) &
                   !is.na(heart.valve$log.lvmi), ]
fit2 <- mjoint(
    formLongFixed = list("grad" = log.grad ~ time + sex + hs,
                         "lvmi" = log.lvmi ~ time + sex),
    formLongRandom = list("grad" = ~ 1 | num,
                          "lvmi" = ~ time | num),
    formSurv = Surv(fuyrs, status) ~ age,
    data = list(hvd, hvd),
    inits = list("gamma" = c(0.11, 1.51, 0.80)),
    timeVar = "time")

AIC(fit2)
BIC(fit2)
}
}
\seealso{
\code{\link[stats]{AIC}} and \code{\link[stats]{BIC}} for the generic methods description.
}
\author{
Alessandro Gasparini (\email{ag475@leicester.ac.uk})
}
\keyword{methods}
