% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsLong.R
\name{initsLong}
\alias{initsLong}
\title{Internal function for generating initial parameters for the longitudinal
sub-model}
\usage{
initsLong(lfit, inits, l, z, K, p, tol.em, verbose)
}
\description{
Internal function for generating initial parameters for the longitudinal
sub-model
}
\keyword{internal}

