% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsSurv.R
\name{initsSurv_balanced}
\alias{initsSurv_balanced}
\title{Internal function for generating initial parameters for the survival
sub-model when the data are *balanced*, i.e. each longitudinal outcome is
measured at the same time within a patient}
\usage{
initsSurv_balanced(data, lfit, survdat2, formSurv, id, timeVar, K, q)
}
\description{
Internal function for generating initial parameters for the survival
sub-model when the data are *balanced*, i.e. each longitudinal outcome is
measured at the same time within a patient
}
\keyword{internal}

