% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsSurv.R
\name{initsSurv_unbalanced}
\alias{initsSurv_unbalanced}
\title{Internal function for generating initial parameters for the survival
sub-model when the data are *unbalanced*, i.e. longitudinal outcomes can be
measured at different times and frequencies within a patient}
\usage{
initsSurv_unbalanced(sfit, K, q)
}
\description{
Internal function for generating initial parameters for the survival
sub-model when the data are *unbalanced*, i.e. longitudinal outcomes can be
measured at different times and frequencies within a patient
}
\keyword{internal}

