% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jLike.R
\name{jLike}
\alias{jLike}
\title{Calculate the log-likelihood and posterior random effects at the maximizer}
\usage{
jLike(theta, l, t, z, nMC)
}
\description{
Calculate the log-likelihood and posterior random effects at the maximizer
}
\keyword{internal}

