% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joineRML.R
\docType{package}
\name{joineRML-package}
\alias{joineRML-package}
\title{joineRML: Joint Modelling of Multivariate Longitudinal Data and Time-to-Event Outcomes}
\description{
Fits the joint model proposed by Henderson and colleagues (2000) \doi{10.1093/biostatistics/1.4.465}, but extended to the case of multiple continuous longitudinal measures. The time-to-event data is modelled using a Cox proportional hazards regression model with time-varying covariates. The multiple longitudinal outcomes are modelled using a multivariate version of the Laird and Ware linear mixed model. The association is captured by a multivariate latent Gaussian process. The model is estimated using a Monte Carlo Expectation Maximization algorithm. This project was funded by the Medical Research Council (Grant number MR/M013227/1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/graemeleehickey/joineRML}
  \item Report bugs at \url{https://github.com/graemeleehickey/joineRML/issues}
}

}
\author{
\strong{Maintainer}: Graeme L. Hickey \email{graemeleehickey@gmail.com} (\href{https://orcid.org/0000-0002-4989-0054}{ORCID})

Authors:
\itemize{
  \item Pete Philipson \email{peter.philipson1@newcastle.ac.uk} (\href{https://orcid.org/0000-0001-7846-0208}{ORCID})
  \item Ruwanthi Kolamunnage-Dona \email{kdrr@liverpool.ac.uk} (\href{https://orcid.org/0000-0003-3886-6208}{ORCID})
  \item Alessandro Gasparini \email{alessandro.gasparini@ki.se} (\href{https://orcid.org/0000-0002-8319-7624}{ORCID})
}

Other contributors:
\itemize{
  \item Andrea Jorgensen \email{aljorgen@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-6977-9337}{ORCID}) [contributor]
  \item Paula Williamson \email{p.r.williamson@liverpool.ac.uk} (\href{https://orcid.org/0000-0001-9802-6636}{ORCID}) [contributor]
  \item Dimitris Rizopoulos \email{d.rizopoulos@erasmusmc.nl} (data/renal.rda, R/hessian.R, R/vcov.R) [contributor, data contributor]
  \item Medical Research Council (Grant number: MR/M013227/1) [funder]
}

}
\keyword{internal}
