% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambdaUpdate}
\alias{lambdaUpdate}
\title{Updates of lambda0 (E-step and M-step)}
\usage{
lambdaUpdate(b_, imat_, zt_, pb_, v_, gam, gam_vec, q, nev, h_)
}
\description{
Updates of lambda0 (E-step and M-step)
}
\keyword{internal}

