% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjoint.object.R
\docType{data}
\name{mjoint.object}
\alias{mjoint.object}
\title{Fitted \code{mjoint} object}
\format{An object of class \code{NULL} of length 0.}
\usage{
mjoint.object
}
\value{
A list with the following components. \describe{

  \item{coefficients}{a list with the estimated coefficients. The components
  of this list are: \describe{

  \item{beta}{the vector of fixed effects for the linear mixed effects
  sub-model.}

  \item{D}{the variance-covariance matrix of the random effects.}

  \item{sigma2}{the measurement error standard deviations for the linear
  mixed effects sub-model.}

  \item{haz}{the estimated baseline hazard values for each unique failure
  time.}

  \item{gamma}{the vector of baseline covariates for the survival model and
  the latent association coefficient parameter estimates.}

  }}

  \item{history}{a matrix with parameter estimates at each iteration of the
  MCEM algorithm.}

  \item{nMC.hx}{a vector with the number of Monte Carlo samples for each MCEM
  algorithm iteration.}

  \item{formLongFixed}{a list of formulae for the fixed effects component of
  each longitudinal outcome.}

  \item{formLongRandom}{a list of formulae for the fixed effects component of
  each longitudinal outcome. The length of the list will be equal to
  \code{formLongFixed}.}

  \item{formSurv}{a formula specifying the proportional hazards regression
  model (not including the latent association structure).}

  \item{data}{a list of data.frames for each longitudinal outcome.}

  \item{survData}{a data.frame of the time-to-event dataset.}

  \item{timeVar}{a character string denoting the column name for time in
  \code{data}.}

  \item{id}{a character string denoting the column name for subject IDs in
  \code{data} and \code{survData}.}

  \item{dims}{a list giving the dimensions of model parameters with
  components: \describe{

  \item{p}{a vector of the number of fixed effects for each longitudinal
  outcome.}

  \item{r}{a vector of the number of random effects for each longitudinal
  outcome.}

  \item{K}{an integer of the number of different longitudinal outcome types.}

  \item{q}{an integer of the number of baseline covariates in the
  time-to-event sub-model.}

  \item{n}{an integer of the total number of subjects in the study.}

  \item{nk}{a vector of the number of measurements for each longitudinal
  outcome.}

  }}

  \item{sfit}{an object of class \code{coxph} for the separate time-to-event
  model fit. See \code{\link[survival]{coxph}} for details.}

  \item{lfit}{a list of objects each of class \code{lme} from fitting
  separate linear mixed effects models; one per each longitudinal outcome
  type. See \code{\link[nlme]{lme}} for details.}

  \item{log.lik0}{the log-likelihood from separate sub-model fits.}

  \item{log.lik}{the log-likelihood from the joint model fit.}

  \item{control}{a list of control parameters used in the estimation of the
  joint model. See \code{\link{mjoint}} for details.}

  \item{finalnMC}{the final number of Monte Carlo samples required prior to
  convergence.}

  \item{SE.approx}{if \code{se.approx=TRUE} in the call to \code{mjoint}, a
  vector of standard errors approximated by the empirical information matrix
  are reported. See \code{\link{mjoint}} for details.}

  \item{Eb}{a matrix with the estimated random effects values for each
  subject.}

  \item{Vb}{an array with the estimated variance-covariance matrices for the
  random effects values for each subject.}

  \item{call}{the matched call.}

  \item{conv}{logical: did the MCEM algorithm converge within the specified
  maximum number of iterations?}

  \item{comp.time}{an object of class \code{difftime} that reports the time
  taken for model fitting.}

  }
}
\description{
An object returned by the \code{mjoint} function, inheriting from class
\code{mjoint} and representing a fitted joint model for multivariate
longitudinal and time-to-event data. Objects of this class have methods for
the generic functions \code{coef}, \code{logLik}, \code{plot}, \code{print},
\code{ranef}, \code{summary}, \code{extractAIC}, and \code{getVarCov}.
}
\author{
Graeme L. Hickey (\email{graeme.hickey@liverpool.ac.uk})
}
\keyword{multivariate}
\keyword{survival}

