% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlme.R
\name{mvlme}
\alias{mvlme}
\title{Internal function for performing a single iteration of the EM algorithm for a
multivariate linear mixed model with possibly unabalnced data}
\usage{
mvlme(thetaLong, l, z, tol.em, verbose)
}
\description{
Internal function for performing a single iteration of the EM algorithm for a
multivariate linear mixed model with possibly unabalnced data
}
\keyword{internal}

