% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{Antithetic simulation of MVN random variables}
\usage{
mvrnormArma(n, mu, sigma)
}
\description{
Antithetic simulation of MVN random variables
}
\keyword{internal}

