% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearPD.R
\name{nearPD}
\alias{nearPD}
\title{Compute the nearest positive definite matrix to an approximate one}
\usage{
nearPD(M, eig.tol = 1e-06, conv.tol = 1e-07, posd.tol = 1e-08,
  maxits = 100)
}
\description{
Compute the nearest positive definite matrix to an approximate one
}
\keyword{internal}

