% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bootSE.R
\name{print.bootSE}
\alias{print.bootSE}
\title{Print function for bootSE object}
\usage{
\method{print}{bootSE}(x, digits = max(4, getOption("digits") - 4), ...)
}
\description{
Print function for bootSE object
}
\keyword{internal}

