% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mjoint.R
\name{print.mjoint}
\alias{print.mjoint}
\title{Print function for mjoint object}
\usage{
\method{print}{mjoint}(x, digits = max(4, getOption("digits") - 4), ...)
}
\description{
Print function for mjoint object
}
\keyword{internal}

