% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mjoint.R
\name{print.summary.mjoint}
\alias{print.summary.mjoint}
\title{Print tidy summary of \code{mjoint} object}
\usage{
\method{print}{summary.mjoint}(x, digits = max(4, getOption("digits") - 4),
  ...)
}
\description{
Print tidy summary of \code{mjoint} object
}
\keyword{internal}

