% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepEM.R
\name{stepEM}
\alias{stepEM}
\title{Internal function for performing a single iteration of the MCEM algorithm}
\usage{
stepEM(theta, l, t, z, nMC, verbose, approxInfo, ll)
}
\description{
Internal function for performing a single iteration of the MCEM algorithm
}
\keyword{internal}

