% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepEM.R
\name{stepEM}
\alias{stepEM}
\title{Internal function for performing a single iteration of the MCEM algorithm}
\usage{
stepEM(theta, l, t, z, nMC, verbose, gammaOpt, postRE, se.approx)
}
\description{
Also takes logical arguments \code{ll} and \code{se.approx}, which calculates
the log-likelihood (and posterior mean and variance of the random effects)
and approximate standard errors, respectively. When either of these arguments
are \code{TRUE}, the function does not return the maaximizer from the MCEM
algorithm iteration, and instead reports the called for post-fit statistics.
}
\keyword{internal}

